fix access control

+++ src/roles/MembersListLocal.js
@@ -3,15 +3,13 @@ import React, {useState,useEffect} from "react"
 import {Row, Card, Col, Badge,Button} from "react-bootstrap"
 import {RiDeleteBin7Line} from "react-icons/ri"
 import {AiOutlineDatabase} from "react-icons/ai"
-import {GrUserAdmin} from "react-icons/gr"
 import {WOQLTable} from '@terminusdb/terminusdb-react-table'
-import {getUsersListConfig,getUsersDatabaseListConfig} from "../ViewConfig"
+import {getUsersListConfigLocal} from "../ViewConfig"
 import {AccessControlHook} from "../hooks/AccessControlHook"
 import {RoleListModal} from "./RoleList"
-import {formatCell} from "./formatData"
 import {UserDatabasesListLocal} from "./UserDatabasesListLocal"
-import { RevokeCapability } from "./RevokeCapability"
-import { AddUserCapabilityModal } from "./AddUserCapabilityModal"
+import {RevokeCapability } from "./RevokeCapability"
+import {AddUserCapabilityModal } from "./AddUserCapabilityModal"
 
 export const MembersListLocal = ({organizationInfo,currentUser,accessControlDashboard,options}) => {  
     if(!accessControlDashboard)return ""
@@ -29,9 +27,7 @@ export const MembersListLocal = ({organizationInfo,currentUser,accessControlDash
     const team = organizationInfo.name 
     const teamId = organizationInfo["@id"] 
 
-    const {getOrgUsersLocal,orgUsers,createUserRole,
-          updateUserRole,
-          loading,
+    const {getOrgUsersLocal,orgUsers,loading,
           errorMessage,
           successMessage} =  AccessControlHook(accessControlDashboard)
     
@@ -45,35 +41,8 @@ export const MembersListLocal = ({organizationInfo,currentUser,accessControlDash
     }
 
     const orgUserArr = Array.isArray(orgUsers) ? orgUsers : []
-    let rowCount=orgUserArr.length  
+    let rowCount=orgUserArr.length 
 
-    //if not capability I create the role
-    const changeUserRoleForScope= (currentSelected)=>{
-        setCurrentRoleToUpdate(currentSelected)
-        setShow(true)
-    }
-
-    const revokePermission =()=>{
-        
-    }
-
-    // maybe move in hook
-    const changeUserRole = (role) =>{
-        //alert(role)
-        if(currentRoleToUpdate.capability){
-            updateUserRole(team,currentRoleToUpdate.userid,currentRoleToUpdate.capability,role,currentRoleToUpdate.scope).then(()=>{
-                if(!errorMessage){
-                    setShow(false)
-                } 
-            })
-        }else{
-            createUserRole(team,currentRoleToUpdate.userid,role,currentRoleToUpdate.scope).then(()=>{
-                if(!errorMessage){
-                    setShow(false)
-                }
-            })
-        }
-    }
 
     const deleteUserItem = (selectedRow)=>{
         setSelectTeamRow(selectedRow)
@@ -87,44 +56,27 @@ export const MembersListLocal = ({organizationInfo,currentUser,accessControlDash
     
     //maybe an utilityfunction woqlClient
     function getActionButtons (cell) {       
-        const currentSelected = cell.row.original//formatCell(cell,"TEAM" , team)
-       // if(!accessControlDashboard.isAdmin() || currentSelected.email === localUser.email )return <span className="d-flex"></span>
+        const currentSelected = cell.row.original
+        if(!accessControlDashboard.isAdmin())return <span className="d-flex"></span>
         return <span className="d-flex">          
             {options.interface.memberList.showDatabase && 
             <Button variant="success" size="sm"   title={`show user dataproducts role`} onClick={() => getUserDatabaseList(currentSelected)}>
                 <AiOutlineDatabase/> 
             </Button>}
             {options.interface.memberList.delete &&      
-                <Button variant="danger" size="sm" className="ml-5" title={`delete ${currentSelected.username}`} onClick={() => deleteUserItem(currentSelected)}>              
-                    <RiDeleteBin7Line/> 
-                </Button>
+            <Button variant="danger" size="sm" className="ml-5" title={`delete ${currentSelected.username}`} onClick={() => deleteUserItem(currentSelected)}>              
+                <RiDeleteBin7Line/> 
+            </Button>
             }
-        </span>
-    }
-
-    function getPicture (cell) {
-        const picture = cell.row.original["picture"] ? cell.row.original["picture"]['@value'] : undefined
-        if (!picture) return ""
-        return <img src={picture} 
-                    alt={"Profile"}
-                    className="nav__main__profile__img mr-4"
-                    width="50"/>
-    }
-
-    function formatRoles (cell) {
-        const rolesList = cell.row.original["role"]
-     
-        return rolesList.map(item=><p>{item}</p>
-        )
+            </span>
     }
 
     const propsObj = {show, setShow, team:team,loading,
-        clickButton:changeUserRole,
         errorMessage,
         successMessage}
     
     
-    const tableConfig = getUsersListConfig(10, getActionButtons,getPicture,formatRoles)
+    const tableConfig = getUsersListConfigLocal(10, getActionButtons)
     
     if(loading){
         return  <Row className="mr-5 ml-2">
@@ -161,10 +113,10 @@ export const MembersListLocal = ({organizationInfo,currentUser,accessControlDash
                         </h6>
                     </Col>
                     <Col >
-                            <button onClick={()=>setShowAdd(true)} style={{maxWidth:"200px"}} title="Create New Role"
-                                    type="button" className="btn-new-data-product mr-1 pt-2 pb-2 pr-4 pl-4 btn btn-sm btn btn-info">
-                                       Add User to {team} Organization
-                             </button>
+                        <button onClick={()=>setShowAdd(true)} style={{maxWidth:"200px"}} title="Create New Role"
+                            type="button" className="btn-new-data-product mr-1 pt-2 pb-2 pr-4 pl-4 btn btn-sm btn btn-info">
+                            Add User to {team} Organization
+                        </button>
                     </Col>
                     </Row>
                 </Card.Header>
